Struct google_people1::api::PeopleMethods

pub struct PeopleMethods<'a, C> where
    C: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on people resources. It is not used directly, but through the PeopleService hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_people1 as people1;
 
use std::default::Default;
use oauth2;
use people1::PeopleService;
 
let secret: oauth2::ApplicationSecret = Default::default();
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = PeopleService::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `batch_create_contacts(...)`, `batch_delete_contacts(...)`, `batch_update_contacts(...)`, `connections_list(...)`, `create_contact(...)`, `delete_contact(...)`, `delete_contact_photo(...)`, `get(...)`, `get_batch_get(...)`, `list_directory_people(...)`, `search_contacts(...)`, `search_directory_people(...)`, `update_contact(...)` and `update_contact_photo(...)`
// to build up your call.
let rb = hub.people();

Implementations

impl<'a, C> PeopleMethods<'a, C>

pub fn connections_list(
&self,
resource_name: &str
) -> PeopleConnectionListCall<'a, C>

Create a builder to help you perform the following task:

Provides a list of the authenticated user's contacts. The request returns a 400 error if personFields is not specified. The request returns a 410 error if sync_token is specified and is expired. Sync tokens expire after 7 days to prevent data drift between clients and the server. To handle a sync token expired error, a request should be sent without sync_token to get all contacts.

Arguments

resourceName - Required. The resource name to return connections for. Only people/me is valid.

pub fn batch_create_contacts(
&self,
request: BatchCreateContactsRequest
) -> PeopleBatchCreateContactCall<'a, C>

Create a builder to help you perform the following task:

Create a batch of new contacts and return the PersonResponses for the newly created contacts. Limited to 10 parallel requests per user.

Arguments

request - No description provided.

pub fn batch_delete_contacts(
&self,
request: BatchDeleteContactsRequest
) -> PeopleBatchDeleteContactCall<'a, C>

Create a builder to help you perform the following task:

Delete a batch of contacts. Any non-contact data will not be deleted. Limited to 10 parallel requests per user.

Arguments

request - No description provided.

pub fn batch_update_contacts(
&self,
request: BatchUpdateContactsRequest
) -> PeopleBatchUpdateContactCall<'a, C>

Create a builder to help you perform the following task:

Update a batch of contacts and return a map of resource names to PersonResponses for the updated contacts. Limited to 10 parallel requests per user.

Arguments

request - No description provided.

pub fn create_contact(&self, request: Person) -> PeopleCreateContactCall<'a, C>

Create a builder to help you perform the following task:

Create a new contact and return the person resource for that contact. The request returns a 400 error if more than one field is specified on a field that is a singleton for contact sources: * biographies * birthdays * genders * names

Arguments

request - No description provided.

pub fn delete_contact(
&self,
resource_name: &str
) -> PeopleDeleteContactCall<'a, C>

Create a builder to help you perform the following task:

Delete a contact person. Any non-contact data will not be deleted.

Arguments

resourceName - Required. The resource name of the contact to delete.

pub fn delete_contact_photo(
&self,
resource_name: &str
) -> PeopleDeleteContactPhotoCall<'a, C>

Create a builder to help you perform the following task:

Delete a contact's photo.

Arguments

resourceName - Required. The resource name of the contact whose photo will be deleted.

pub fn get(&self, resource_name: &str) -> PeopleGetCall<'a, C>

Create a builder to help you perform the following task:

Provides information about a person by specifying a resource name. Use people/me to indicate the authenticated user. The request returns a 400 error if 'personFields' is not specified.

Arguments

resourceName - Required. The resource name of the person to provide information about. - To get information about the authenticated user, specify people/me. - To get information about a google account, specify people/{account_id}. - To get information about a contact, specify the resource name that identifies the contact as returned by people.connections.list.

pub fn get_batch_get(&self) -> PeopleGetBatchGetCall<'a, C>

Create a builder to help you perform the following task:

Provides information about a list of specific people by specifying a list of requested resource names. Use people/me to indicate the authenticated user. The request returns a 400 error if 'personFields' is not specified.

pub fn list_directory_people(&self) -> PeopleListDirectoryPeopleCall<'a, C>

Create a builder to help you perform the following task:

Provides a list of domain profiles and domain contacts in the authenticated user's domain directory.

pub fn search_contacts(&self) -> PeopleSearchContactCall<'a, C>

Create a builder to help you perform the following task:

Provides a list of contacts in the authenticated user's grouped contacts that matches the search query. The query matches on a contact's names, nickNames, emailAddresses, phoneNumbers, and organizations fields that are from the CONTACT" source.

pub fn search_directory_people(&self) -> PeopleSearchDirectoryPeopleCall<'a, C>

Create a builder to help you perform the following task:

Provides a list of domain profiles and domain contacts in the authenticated user's domain directory that match the search query.

pub fn update_contact(
&self,
request: Person,
resource_name: &str
) -> PeopleUpdateContactCall<'a, C>

Create a builder to help you perform the following task:

Update contact data for an existing contact person. Any non-contact data will not be modified. Any non-contact data in the person to update will be ignored. All fields specified in the update_mask will be replaced. The server returns a 400 error if person.metadata.sources is not specified for the contact to be updated or if there is no contact source. The server returns a 400 error with reason "failedPrecondition" if person.metadata.sources.etag is different than the contact's etag, which indicates the contact has changed since its data was read. Clients should get the latest person and merge their updates into the latest person. The server returns a 400 error if memberships are being updated and there are no contact group memberships specified on the person. The server returns a 400 error if more than one field is specified on a field that is a singleton for contact sources: * biographies * birthdays * genders * names

Arguments

request - No description provided.

resourceName - The resource name for the person, assigned by the server. An ASCII string with a max length of 27 characters, in the form of people/{person_id}.

pub fn update_contact_photo(
&self,
request: UpdateContactPhotoRequest,
resource_name: &str
) -> PeopleUpdateContactPhotoCall<'a, C>

Create a builder to help you perform the following task:

Update a contact's photo.

Arguments

request - No description provided.

resourceName - Required. Person resource name