Struct google_people1::api::Person

pub struct Person {
    pub addresses: Option<Vec<Address>>,
    pub age_range: Option<String>,
    pub age_ranges: Option<Vec<AgeRangeType>>,
    pub biographies: Option<Vec<Biography>>,
    pub birthdays: Option<Vec<Birthday>>,
    pub bragging_rights: Option<Vec<BraggingRights>>,
    pub calendar_urls: Option<Vec<CalendarUrl>>,
    pub client_data: Option<Vec<ClientData>>,
    pub cover_photos: Option<Vec<CoverPhoto>>,
    pub email_addresses: Option<Vec<EmailAddress>>,
    pub etag: Option<String>,
    pub events: Option<Vec<Event>>,
    pub external_ids: Option<Vec<ExternalId>>,
    pub file_ases: Option<Vec<FileAs>>,
    pub genders: Option<Vec<Gender>>,
    pub im_clients: Option<Vec<ImClient>>,
    pub interests: Option<Vec<Interest>>,
    pub locales: Option<Vec<Locale>>,
    pub locations: Option<Vec<Location>>,
    pub memberships: Option<Vec<Membership>>,
    pub metadata: Option<PersonMetadata>,
    pub misc_keywords: Option<Vec<MiscKeyword>>,
    pub names: Option<Vec<Name>>,
    pub nicknames: Option<Vec<Nickname>>,
    pub occupations: Option<Vec<Occupation>>,
    pub organizations: Option<Vec<Organization>>,
    pub phone_numbers: Option<Vec<PhoneNumber>>,
    pub photos: Option<Vec<Photo>>,
    pub relations: Option<Vec<Relation>>,
    pub relationship_interests: Option<Vec<RelationshipInterest>>,
    pub relationship_statuses: Option<Vec<RelationshipStatus>>,
    pub residences: Option<Vec<Residence>>,
    pub resource_name: Option<String>,
    pub sip_addresses: Option<Vec<SipAddress>>,
    pub skills: Option<Vec<Skill>>,
    pub taglines: Option<Vec<Tagline>>,
    pub urls: Option<Vec<Url>>,
    pub user_defined: Option<Vec<UserDefined>>,
}

Information about a person merged from various data sources such as the authenticated user's contacts and profile data. Most fields can have multiple items. The items in a field have no guaranteed order, but each non-empty field is guaranteed to have exactly one field with metadata.primary set to true.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

copy other contact to my contacts group other contacts (response)

create contact people (request|response)

get people (response)

update contact people (request|response)

Fields
addresses: Option<Vec<Address>>

The person's street addresses.

age_range: Option<String>

Output only. DEPRECATED (Please use person.ageRanges instead) The person's age range.

age_ranges: Option<Vec<AgeRangeType>>

Output only. The person's age ranges.

biographies: Option<Vec<Biography>>

The person's biographies. This field is a singleton for contact sources.

birthdays: Option<Vec<Birthday>>

The person's birthdays. This field is a singleton for contact sources.

bragging_rights: Option<Vec<BraggingRights>>

DEPRECATED: No data will be returned The person's bragging rights.

calendar_urls: Option<Vec<CalendarUrl>>

The person's calendar URLs.

client_data: Option<Vec<ClientData>>

The person's client data.

cover_photos: Option<Vec<CoverPhoto>>

Output only. The person's cover photos.

email_addresses: Option<Vec<EmailAddress>>

The person's email addresses.

etag: Option<String>

The HTTP entity tag of the resource. Used for web cache validation.

events: Option<Vec<Event>>

The person's events.

external_ids: Option<Vec<ExternalId>>

The person's external IDs.

file_ases: Option<Vec<FileAs>>

The person's file-ases.

genders: Option<Vec<Gender>>

The person's genders. This field is a singleton for contact sources.

im_clients: Option<Vec<ImClient>>

The person's instant messaging clients.

interests: Option<Vec<Interest>>

The person's interests.

locales: Option<Vec<Locale>>

The person's locale preferences.

locations: Option<Vec<Location>>

The person's locations.

memberships: Option<Vec<Membership>>

The person's group memberships.

metadata: Option<PersonMetadata>

Output only. Metadata about the person.

misc_keywords: Option<Vec<MiscKeyword>>

The person's miscellaneous keywords.

names: Option<Vec<Name>>

The person's names. This field is a singleton for contact sources.

nicknames: Option<Vec<Nickname>>

The person's nicknames.

occupations: Option<Vec<Occupation>>

The person's occupations.

organizations: Option<Vec<Organization>>

The person's past or current organizations.

phone_numbers: Option<Vec<PhoneNumber>>

The person's phone numbers.

photos: Option<Vec<Photo>>

Output only. The person's photos.

relations: Option<Vec<Relation>>

The person's relations.

relationship_interests: Option<Vec<RelationshipInterest>>

Output only. DEPRECATED: No data will be returned The person's relationship interests.

relationship_statuses: Option<Vec<RelationshipStatus>>

Output only. DEPRECATED: No data will be returned The person's relationship statuses.

residences: Option<Vec<Residence>>

DEPRECATED: (Please use person.locations instead) The person's residences.

resource_name: Option<String>

The resource name for the person, assigned by the server. An ASCII string with a max length of 27 characters, in the form of people/{person_id}.

sip_addresses: Option<Vec<SipAddress>>

The person's SIP addresses.

skills: Option<Vec<Skill>>

The person's skills.

taglines: Option<Vec<Tagline>>

Output only. DEPRECATED: No data will be returned The person's taglines.

urls: Option<Vec<Url>>

The person's associated URLs.

user_defined: Option<Vec<UserDefined>>

The person's user defined data.