Struct google_people1::Name

pub struct Name {
    pub phonetic_middle_name: Option<String>,
    pub honorific_prefix: Option<String>,
    pub phonetic_family_name: Option<String>,
    pub display_name: Option<String>,
    pub middle_name: Option<String>,
    pub phonetic_honorific_prefix: Option<String>,
    pub family_name: Option<String>,
    pub phonetic_given_name: Option<String>,
    pub phonetic_honorific_suffix: Option<String>,
    pub given_name: Option<String>,
    pub honorific_suffix: Option<String>,
    pub metadata: Option<FieldMetadata>,
}

A person's name. If the name is a mononym, the family name is empty.

This type is not used in any activity, and only used as part of another schema.

Fields

phonetic_middle_name

The middle name(s) spelled as they sound.

honorific_prefix

The honorific prefixes, such as Mrs. or Dr.

phonetic_family_name

The family name spelled as it sounds.

display_name

The display name formatted according to the locale specified by the viewer's account or the Accept-Language HTTP header.

middle_name

The middle name(s).

phonetic_honorific_prefix

The honorific prefixes spelled as they sound.

family_name

The family name.

phonetic_given_name

The given name spelled as it sounds.

phonetic_honorific_suffix

The honorific suffixes spelled as they sound.

given_name

The given name.

honorific_suffix

The honorific suffixes, such as Jr.

metadata

Metadata about the name.