Struct google_people1::Organization

pub struct Organization {
    pub formatted_type: Option<String>,
    pub domain: Option<String>,
    pub end_date: Option<Date>,
    pub name: Option<String>,
    pub start_date: Option<Date>,
    pub symbol: Option<String>,
    pub title: Option<String>,
    pub current: Option<bool>,
    pub job_description: Option<String>,
    pub location: Option<String>,
    pub department: Option<String>,
    pub type_: Option<String>,
    pub phonetic_name: Option<String>,
    pub metadata: Option<FieldMetadata>,
}

A person's past or current organization. Overlapping date ranges are permitted.

This type is not used in any activity, and only used as part of another schema.

Fields
formatted_type: Option<String>

The read-only type of the organization translated and formatted in the viewer's account locale or the Accept-Language HTTP header locale.

domain: Option<String>

The domain name associated with the organization; for example, google.com.

end_date: Option<Date>

The end date when the person left the organization.

name: Option<String>

The name of the organization.

start_date: Option<Date>

The start date when the person joined the organization.

symbol: Option<String>

The symbol associated with the organization; for example, a stock ticker symbol, abbreviation, or acronym.

title: Option<String>

The person's job title at the organization.

current: Option<bool>

True if the organization is the person's current organization; false if the organization is a past organization.

job_description: Option<String>

The person's job description at the organization.

location: Option<String>

The location of the organization office the person works at.

department: Option<String>

The person's department at the organization.

type_: Option<String>

The type of the organization. The type can be custom or predefined. Possible values include, but are not limited to, the following: * work * school

phonetic_name: Option<String>

The phonetic name of the organization.

metadata: Option<FieldMetadata>

Metadata about the organization.