Struct google_people1::PeopleMethods

pub struct PeopleMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on people resources. It is not used directly, but through the People hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_people1 as people1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use people1::People;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = People::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `connections_list(...)`, `get(...)` and `get_batch_get(...)`
// to build up your call.
let rb = hub.people();

Methods

impl<'a, C, A> PeopleMethods<'a, C, A>

fn get(&self, resource_name: &str) -> PeopleGetCall<'a, C, A>

Create a builder to help you perform the following task:

Provides information about a person resource for a resource name. Use people/me to indicate the authenticated user.

Arguments

resourceName - The resource name of the person to provide information about. - To get information about the authenticated user, specify people/me. - To get information about any user, specify the resource name that identifies the user, such as the resource names returned by people.connections.list.

fn connections_list(&self, resource_name: &str) -> PeopleConnectionListCall<'a, C, A>

Create a builder to help you perform the following task:

Provides a list of the authenticated user's contacts merged with any linked profiles.

Arguments

resourceName - The resource name to return connections for. Only people/me is valid.

fn get_batch_get(&self) -> PeopleGetBatchGetCall<'a, C, A>

Create a builder to help you perform the following task:

Provides information about a list of specific people by specifying a list of requested resource names. Use people/me to indicate the authenticated user.