Struct google_people1::Person

pub struct Person {
    pub interests: Option<Vec<Interest>>,
    pub bragging_rights: Option<Vec<BraggingRights>>,
    pub addresses: Option<Vec<Address>>,
    pub nicknames: Option<Vec<Nickname>>,
    pub occupations: Option<Vec<Occupation>>,
    pub phone_numbers: Option<Vec<PhoneNumber>>,
    pub age_range: Option<String>,
    pub photos: Option<Vec<Photo>>,
    pub names: Option<Vec<Name>>,
    pub biographies: Option<Vec<Biography>>,
    pub residences: Option<Vec<Residence>>,
    pub resource_name: Option<String>,
    pub relationship_interests: Option<Vec<RelationshipInterest>>,
    pub cover_photos: Option<Vec<CoverPhoto>>,
    pub locales: Option<Vec<Locale>>,
    pub organizations: Option<Vec<Organization>>,
    pub skills: Option<Vec<Skill>>,
    pub birthdays: Option<Vec<Birthday>>,
    pub relations: Option<Vec<Relation>>,
    pub memberships: Option<Vec<Membership>>,
    pub taglines: Option<Vec<Tagline>>,
    pub etag: Option<String>,
    pub urls: Option<Vec<Url>>,
    pub relationship_statuses: Option<Vec<RelationshipStatus>>,
    pub genders: Option<Vec<Gender>>,
    pub email_addresses: Option<Vec<EmailAddress>>,
    pub im_clients: Option<Vec<ImClient>>,
    pub events: Option<Vec<Event>>,
    pub metadata: Option<PersonMetadata>,
}

Information about a person merged from various data sources such as the authenticated user's contacts and profile data. Fields other than IDs, metadata, and group memberships are user-edited. Most fields can have multiple items. The items in a field have no guaranteed order, but each non-empty field is guaranteed to have exactly one field with metadata.primary set to true.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get people (response)

Fields

interests

The person's interests.

bragging_rights

The person's bragging rights.

addresses

The person's street addresses.

nicknames

The person's nicknames.

occupations

The person's occupations.

phone_numbers

The person's phone numbers.

age_range

The person's age range.

photos

The person's photos.

names

The person's names.

biographies

The person's biographies.

residences

The person's residences.

resource_name

The resource name for the person, assigned by the server. An ASCII string with a max length of 27 characters. Always starts with people/.

relationship_interests

The kind of relationship the person is looking for.

cover_photos

The person's cover photos.

locales

The person's locale preferences.

organizations

The person's past or current organizations.

skills

The person's skills.

birthdays

The person's birthdays.

relations

The person's relations.

memberships

The person's group memberships.

taglines

The person's taglines.

etag

The HTTP entity tag of the resource. Used for web cache validation.

urls

The person's associated URLs.

relationship_statuses

The person's relationship statuses.

genders

The person's genders.

email_addresses

The person's email addresses.

im_clients

The person's instant messaging clients.

events

The person's events.

metadata

Metadata about the person.