Struct google_people1::Person

pub struct Person {
    pub interests: Option<Vec<Interest>>,
    pub bragging_rights: Option<Vec<BraggingRights>>,
    pub addresses: Option<Vec<Address>>,
    pub nicknames: Option<Vec<Nickname>>,
    pub occupations: Option<Vec<Occupation>>,
    pub phone_numbers: Option<Vec<PhoneNumber>>,
    pub age_range: Option<String>,
    pub photos: Option<Vec<Photo>>,
    pub names: Option<Vec<Name>>,
    pub biographies: Option<Vec<Biography>>,
    pub residences: Option<Vec<Residence>>,
    pub resource_name: Option<String>,
    pub relationship_interests: Option<Vec<RelationshipInterest>>,
    pub cover_photos: Option<Vec<CoverPhoto>>,
    pub locales: Option<Vec<Locale>>,
    pub organizations: Option<Vec<Organization>>,
    pub skills: Option<Vec<Skill>>,
    pub birthdays: Option<Vec<Birthday>>,
    pub relations: Option<Vec<Relation>>,
    pub memberships: Option<Vec<Membership>>,
    pub taglines: Option<Vec<Tagline>>,
    pub etag: Option<String>,
    pub urls: Option<Vec<Url>>,
    pub relationship_statuses: Option<Vec<RelationshipStatus>>,
    pub genders: Option<Vec<Gender>>,
    pub email_addresses: Option<Vec<EmailAddress>>,
    pub im_clients: Option<Vec<ImClient>>,
    pub events: Option<Vec<Event>>,
    pub metadata: Option<PersonMetadata>,
}

Information about a person merged from various data sources such as the authenticated user's contacts and profile data. Fields other than IDs, metadata, and group memberships are user-edited. Most fields can have multiple items. The items in a field have no guaranteed order, but each non-empty field is guaranteed to have exactly one field with metadata.primary set to true.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get people (response)

Fields
interests: Option<Vec<Interest>>

The person's interests.

bragging_rights: Option<Vec<BraggingRights>>

The person's bragging rights.

addresses: Option<Vec<Address>>

The person's street addresses.

nicknames: Option<Vec<Nickname>>

The person's nicknames.

occupations: Option<Vec<Occupation>>

The person's occupations.

phone_numbers: Option<Vec<PhoneNumber>>

The person's phone numbers.

age_range: Option<String>

The person's age range.

photos: Option<Vec<Photo>>

The person's photos.

names: Option<Vec<Name>>

The person's names.

biographies: Option<Vec<Biography>>

The person's biographies.

residences: Option<Vec<Residence>>

The person's residences.

resource_name: Option<String>

The resource name for the person, assigned by the server. An ASCII string with a max length of 27 characters. Always starts with people/.

relationship_interests: Option<Vec<RelationshipInterest>>

The kind of relationship the person is looking for.

cover_photos: Option<Vec<CoverPhoto>>

The person's cover photos.

locales: Option<Vec<Locale>>

The person's locale preferences.

organizations: Option<Vec<Organization>>

The person's past or current organizations.

skills: Option<Vec<Skill>>

The person's skills.

birthdays: Option<Vec<Birthday>>

The person's birthdays.

relations: Option<Vec<Relation>>

The person's relations.

memberships: Option<Vec<Membership>>

The person's group memberships.

taglines: Option<Vec<Tagline>>

The person's taglines.

etag: Option<String>

The HTTP entity tag of the resource. Used for web cache validation.

urls: Option<Vec<Url>>

The person's associated URLs.

relationship_statuses: Option<Vec<RelationshipStatus>>

The person's relationship statuses.

genders: Option<Vec<Gender>>

The person's genders.

email_addresses: Option<Vec<EmailAddress>>

The person's email addresses.

im_clients: Option<Vec<ImClient>>

The person's instant messaging clients.

events: Option<Vec<Event>>

The person's events.

metadata: Option<PersonMetadata>

Metadata about the person.